requirements.txt
mss>=10.1.0
opencv-python>=4.7.0
numpy>=1.26.0
keyboard>=0.13.0
pygetwindow>=0.0.9
Pillow>=11.0.0
pyautogui>=0.9.0
pyperclip>=1.8.0
requests>=2.25.0
pytz>=2021.1
cryptography>=41.0.0
PyInstaller>=6.0.0